무한루프 오류 수정
