+++ src/index.tsx
@@ -1,24 +1,11 @@
-import React, { useEffect } from "react";
+import React from "react";
 import ReactDOM from "react-dom/client";
+import reportWebVitals from "./reportWebVitals";
 import { BrowserRouter } from "react-router-dom";
 import { HelmetProvider } from "react-helmet-async";
-import useLoginUserStore from "src/stores/useLoginUserStore";
+
 import "./index.css"; //tailwind css
 import App from "./App";
-
-import { getLocalItem } from "./util/storage";
-import User from "./constants/User";
-
-const Root = () => {
-  console.log("index() 시작");
-
-  const { setLoginUser, loginUser } = useLoginUserStore();
-  if (!loginUser) {
-    setLoginUser(JSON.parse(getLocalItem(User.LOGIN_USER))); //로그인 정보 store에 저장
-  }
-
-  return <App />;
-};
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
@@ -27,7 +14,7 @@ root.render(
   // <React.StrictMode>
   <BrowserRouter>
     <HelmetProvider>
-      <Root />
+      <App />
     </HelmetProvider>
   </BrowserRouter>
   // </React.StrictMode>
