+++ src/routes/AuthRoute.tsx
@@ -6,10 +6,6 @@ import Header from "src/components/common/layout/Header";
 import SideBar from "src/components/common/layout/SideBar";
 
 const AuthoRoute = () => {
-  const navigate = useNavigate();
-  const { pathname } = useLocation();
-  //const { user } = useLoginUserStore();
-
   // useEffect(() => {
   //   if (!user) {
   //     alert('로그인 후 이용해주세요');
@@ -18,13 +14,10 @@ const AuthoRoute = () => {
   // }, [navigate, pathname, user]);
 
   return (
-    <div className="flex h-screen w-full flex-col">
-      <Header />
-      <div className="flex flex-1">
-        <SideBar />
-        <div className="flex-1 bg-gray-50 p-6">
-          <Outlet />
-        </div>
+    <div className="flex flex-1">
+      <SideBar />
+      <div className="flex-1 bg-gray-50 p-6">
+        <Outlet />
       </div>
     </div>
   );
